lldpad: implementation of IEEE 802.1Qbg in lldpad, part 2
This is the implementation of the edge control protocol (ECP) and VSI
discovery protocol (VDP) currently being specified in IEEE 802.1Qbg.

This implementation extends the infrastructure defined lldpad to send
and receive ECP frames with a new (yet to be defined) ethertype.
Received frames are validated and analyzed before the content is
handed to the upper layer protocol (ULP, VDP in this case) for
further processing. Frames to be transmitted are compiled from VSI
(guest interface) profiles registered on a interface. Reception and
transmission of ECP frames is controlled by RX and TX state machines,
timeouts are handled timeout functions. The patch still contains a
lot of debug code to allow low-level protocol analysis.

VDP serves as the upper layer protocol (ULP) for TLVs communicated
via the ECP protocol.

For this it registers as a new module in lldpad. The VDP module
supports a station and a bridge role. As a station, new VSI (virtual
station interface) profiles can be registered to the VDP module using
lldptool or libvirt. These profiles are then announced to an adjacent
switch. Transmitted profiles are processed to the desired state by
the VDP station state machine. As a bridge, the VDP module waits for
new profiles received in TLVs by ECP. The received profiles are
processed to the desired state by a VDP bridge state machine.

VDP module parameters are stored in the "vdp" section under the
appropriate interface.

The patch still contains a lot of debug code to allow analysis of VDP
protocol behavior.

Signed-off-by: Jens Osterkamp <jens at linux.vnet.ibm.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>