lldpad: VDP commandline interface
This patch implements the command line interface to control the VDP
module. In station role, it allows to register a new VSI (guest
interface) profile and query its state. In bridge role, it allows to
query the state of all registered profiles.

With lldptool it is possible to set the role from default "station" to
"bridge". The configuration of the role for a port is saved to lldpads
config file.

Signed-off-by: Jens Osterkamp <jens at linux.vnet.ibm.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>