lldpad: do not use macv[tap/lan] interfaces as ports
At startup lldpad walks through all network interfaces in the system
and creates internal data structures for them. Some interfaces as
e.g. vlan and wlan are skipped in the walkthrough, some have to be
treated special (e.g. bond devices).

This patch adds macvtap and macvlan interfaces to the list of devices
that are skipped as we do not want to send out EVB/ECP frames on them.

Signed-off-by: Jens Osterkamp <jens at linux.vnet.ibm.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>