lldpad: fix evb init segfault
If the evb tlv is not enabled then evb_bld_cfg_tlv fails. This is
not a critical failure it only means the TLV has been disabled by the
user.

Allow initialization to continue on ifup so users can dynamically
enable EVB on active devices. And do not post errors when evb_gettlv()
fails. This is expected when the tlv is disabled. Notice the other
case where the TLV is enabled on ifup but is later disabled.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>