new: rework of VDP code towards newer drafts
This is a first rework of the VDP towards the changes in newer drafts.
It contains the following changes:

- bugfix: prevent sending of duplicate keepalives

While checking ECP traffic on the wire a lot of ECP frames being
sent until an ack was received. This fixes it.

- allow transition from UNASSOCIATED to DEASSOCIATED

This is necessary in cases when lldpad has been terminated,
but VMs are still running, to allow these VMs to be deassociated.

- do not remove profile on ifdown

Up to now profiles were deleted and the vdp_data struct removed from the
list on "ifconfig eth? down".
Now profiles are retained during interface down. Processing of profiles will
continue after the interface comes up again. Keepalives are sent as before.

Signed-off-by: Jens Osterkamp <jens@linux.vnet.ibm.com>