prevent "forgotten" ECP frame for vsi type
We have seen cases where an ecp frame for a profile is "forgotten" to be
sent out if the ecp tx state machine is still waiting for an ack for a
previous frame.

The sequence which leads to this is as follows:

1) ecp frame for vsi type 1 is sent out
2) ecp tx state machine is switched to ECP_TX_WAIT_FOR_ACK_FRAME
3) keepalive timer for vsi type 2 times out, vsi type 2 is marked fo
4) ecp tx state machine switches to ECP_TX_REQUEST_PDU
5) ecp frame for vsi type 2 is not sent out
6) vdp ack timer for vsi type 2 times out
7) vsi type 2 is removed

It can take between hours and days until this behaviour can be observed.
This patch fixes it by checking for pending vsi type when switching to
ECP_TX_REQUEST_PDU.

Signed-off-by: Jens Osterkamp <jens@linux.vnet.ibm.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>