lldpad: remove cur_peer and last_peer from DCBx
cur_peer and last_peer are only used in the dcbx codepath
and worse their use in dcbx is buggy when dcbxv1 and dcbxv2
are present in the same lldp frame.

Rather than try to resolve these issues this patch strips
them out ocmpletely. And uses the stored tlv in the dcbx
struct when needed. This resolves the issues and removes
some unneeded code.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>