lldpad: cleanup, fix check in gettlv for null ptr tlvs
Move check for null tlvs data structure in dcbx_gettlv()
above free tlvs. If there are no tlvs there is no reason
to try and free them.

Fortunantly, the free tlvs routines also check for a
null ptr so no harm is caused by doing this. But this
cleanup makes it read a bit better.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>