lldpad: dcbx, dropped ifdown messages breaks module
Because link events can occasionally be lost we have made
the lldpad core handle this case assuming the modules will
initialize correctly when the ifup() routine is called even
when the corresponding ifdown() was not. The dcbx module
however was broke in this case and the result was an
uninitialized port. This patch fixes this by moving the goto
up the routine to initialize the interface even if ifdown
was skipped.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>