lldpad: DCB, remove iSCSI tc qdisc and filters
This removes creating the multiq qdisc and iSCSI filter when
the iSCSI APP TLV is set. The multq qdisc was being created
and a filter to map iSCSI traffic to the queue number that is
being advertised as the app priority for iSCSI was setup.
But this implementation _only_ works on ixgbe drivers before
kernel 2.6.38 where this queue is actually mapped to the
correct traffic class. Probably worse is it breaks other
CNAs that use the kernel API select_queue() or skb->priority
to traffic class mappings to steer traffic.

Additionally, even if the host is not running iSCSI at all
the qdisc and filter are added if the switch advertises
the iSCSI APP TLV. By adding the qdisc we adversely effect
FCoE performance.

This patch removes the adding of the qdisc and filters
resolving the issues above.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>