reduce number of select timeouts for ECP
A problem with the number of select timeouts has been reported in Redhat
bugzilla #694639 (https://bugzilla.redhat.com/show_bug.cgi?id=694639) and
by Will Cohen on the open-lldp mailing list.

The number of select timeouts for any running applications can be measured
with 'stap -c "sleep 10" timeout.stp'.

The high number of select timeouts was caused by the timers used in the ECP
code to control the timeout until when an ack has to be received and to
determine wether there is any work pending (e.g. changed profiles).

This patch changes the timers from calling them at a regular interval to a
demand based model. For this it is necessary to setup a short-running timer
whenever work has been scheduled. This way the execution path returns to the
event loop much more often leaving other modules a chance to get work done.

Signed-off-by: Jens Osterkamp <jens@linux.vnet.ibm.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>