lldpad: pass obuf length to cmd handlers
Use safe versions of string copy routines. Arguably in some
cases the input is already validated and guarenteed to be
less then the buffer size. However using safe versions in
these cases is OK and removes any errors or missed cases
during review. Or more likely any possiblitiy of a future
change introducing a potentially unvalidated copy.

To do this the length needed to be passed into the command
handlers.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>