lldpad: add -c modify tlv option to commands
Add a -c option to modify TLV query commands to indicate that
the command pertains to the configuration of the TLV instead of
the contents of the local or remote LLDP MIB.  Supplying the
-c argument by itself will result in all of the configuration
elements for the specified TLV to be returned.

This patch also updates the man pages with information about
the -c option.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>