lldpad: delete app tlv entry feature
The patch provides the ability to delete application
TLV configuration entries.  A -d option is added to
lldptool to specify deletion.  Currently, -d will only
do anything useful for the App TLV.  It can be extended
in the future to work for other TLVs which take multiple
values.  Also, a limit on the number of Application entries
allowed to be configured is implemented.  Current value is
32 entries.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>