lldpad: Force changes made from dormant Timer to be seen on the wire
In the case where lldpad is not connected to a peer the dormant timer
will expire and APP values will be added to the TLV and kernel. This
timer is used to avoid reprogramming the device immediately after
ifup only to promptly receive the peers TLV and be forced to reprogram
again.

However currently the tx'd TLV does not get updated until the normal
TX state machine sends a new TLV. It would be best to remove this
delay and send the TLV immediately. This patch kicks the state
machine to let it know APP values have changed.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>