Change the client interface to use an abstract namespace socket.
This is useful in the initrd stage as it does not require
the addition of directories and files to the filesystem just
for the purpose of supporting the client interface.

Passes sender credentials to verify that only root is able to
send control messages.

This patch removes all supporting code for the previous socket
style.

Note: other clients of lldpad also will need to be modified to
utilitize this new socket type (e.g. fcoemon in fcoe-utils).

This patch was original supplied by Hannes Reinecke. His signed
off line was dropped because some changes were required to get
this to apply against latest source.

Suggested-by: Hannes Reinecke <hare@suse.de>
Signed-off-by: Eric Multanen <eric.w.multanen@intel.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>