dcbtool: return different error codes for invalid device and IEEE
dcbtool currently returns the same error code for an invalid
device name and a device in IEEE mode. This results in
the tool printing an output that gives an error about IEEE
mode in both cases.

This patch separates the two cases and additionally allows
sets and gets to occur. Oper mode calls should still fail
device in IEEE mode errors. This is more inline with the
current lldptool behavior and supports users configuring
CEE mode before switching DCBX mode.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>