lldpad: dcbx mode needs to be set before set_state() cmd
lldp_dcbx issues a set_state() command before setting the dcbx
mode. In almost all cases this is OK because we don't use dcbx
mode unless the driver is already setup this way or we received
a CEE DCBX frame which already set the mode.

The corner case being when the user manually forces CEE mode
in this case the driver may be in IEEE mode and we need to
set the mode before issueing CEE commands otherwise the driver
may be expecting IEEE cmds or the driver and lldpad can become
out of sync.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>