lldpad: Fix logging to no longer use message IDs
lldpad's logging was overloading the first parameter to the log_message
function as both a logging level and a message ID. There are really not
many message IDs in use, so resolve the conflict by using only log level.
This also adds a proper printf-attribute to the logging function so
that arguments can be checked and resolve those as well. A few messages
that were changed now have more useful content.

Signed-off-by: Mark Rustad <mark.d.rustad@intel.com>
Tested-by: Ross Brattain <ross.b.brattain@intel.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>