lldpad: Introduce support for multiple agents per port
This patch introduces support for multiple agents per port.
Each agent sends and receives frames to and from different group mac
addresses. Right now these are nearest bridge, nearest customer bridge and
nearest non-twoport mac-relaying bridge.

It splits out most of the agent specific variables from the port structure
to an agent structure. This requires changes in the rx and tx state
machines.

With the changes for the multi-agent support, adminStatus has moved from the
port to the agent. As port->adminStatus has been used to control the VDP
state machines as well, a new mechanism was needed here.
The VDP module does not have agent support, so port->adminStatus is replaced
by the vdp->enableTx flag. If this is set to yes by lldptool, VDP will be
active, and will be inactive if set to no.

JF: fixed some minor nit issues and a few real bugs as well as
    some style items.

Signed-off-by: Jens Osterkamp <jens@linux.vnet.ibm.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>