lldpad: Introduce changed config file interface
This patch modifies all interface functions to get and set config file
values to use the agent type. The agent type will determine in which section
of the config file the values are located. The result is that now there are 3
sections in lldpad.conf for LLDP:

	- "lldp" or "nearest_bridge" (if newly created)
	- "nearest_customer_bridge"
	- "nearest_nontpmr_bridge"

So lldpad.conf will look similar like

	lldp {
		...
	}
	nearest_customer_bridge {
		...
	}
	nearest_nontmpr_bridge {
		...
	}

It also introduces a new lldptool parameter -g <agent>. Valid values for
<agent> are

	- "nearest_bridge" or "nb"
	- "nearest_customer_bridge" or "ncb"
	- "neareast_nontpmr_bridge" or "nntpmrb"

E.g. to enable the evbCfg TLV for the agent "nearest customer bridge" one
would use

	"lldptool -T -i eth8 -g ncb -V evbCfg -c enableTx=yes"

If no -g parameter is specified, this is assumed to remain "nearest bridge"
to preserve the existing behaviour.

It modifies the lldp_mod_* functions to supply the agent where appropriate.

JF: updated to default to nearest bridge if '-g' is not specified.

Signed-off-by: Jens Osterkamp <jens@linux.vnet.ibm.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>