lldpad: changes to the lldp command ctrl interface
This patch bumps the msg version number to reflect the addition of the agent
type to the clif command message.

It adds a check for the CLIF_MSG_VERSION to the lldp clif command processing
which will output an informated "deprecated version" message if the version
is not current. This warns users about lack of agent support.

Suggested-by: Jens Osterkamp <jens@linux.vnet.ibm.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>