lldpad: remain in legacy mode after restart
Some DCBX peers have not behaved well when lldpad is restarted
and begins with IEEE DCBX before detecting that a peer is using
CEE. This patch will store the active legacy DCBX mode in the
shared memory segment and use that info to start up directly in
the legacy mode in cases where lldpad is restarted. This setting
is not persistent across reboots.  To allow a user to restore the
automatic mode where IEEE DCBX is initially attempted, an ability
to 'reset' the IEEE-DCBX 'mode' is provided.

  lldptool -Ti eth1 -V IEEE-DCBX -c mode=reset

Upon query, IEEE-DCBX 'mode' will either respond with 'auto' or
the legacy DCBX mode (e.g. 'CEE').

  lldptool -ti eth1 -V IEEE-DCBX -c mode

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>
Signed-off-by: Eric Multanen <eric.w.multanen@intel.com>
Tested-by: Ross Brattain <ross.b.brattain@intel.com>