lldpad: Fix inconsistent calls to get/set configuration values
Calls to various get/set config functions are being made inconsistently.
Some places pass a pointer to an int and other places pass a pointer
to a long. Resolve to use the former and introduce a transparent
union to prevent this and other kinds of errors from creeping back in.
This also allows open-lldp to build with either 1.3.x or 1.4.x versions
of libconfig by accommodating the change from long to int for the int
operations.

V2: Corrects a mistake that could result in freeing an uninitialized
    pointer.
V3: Handle both 1.3.x and 1.4.x versions of libconfig.
V4: Correct prototypes to avoid casting issue at higher warning levels.
V5: Resolve conflicts with recent multi-agent patches.
V6: Resolve conflicts with recent changes.

Signed-off-by: Mark Rustad <mark.d.rustad@intel.com>