lldpad: fix somewhat broken DCBX mapping routine
The DCBX mapping routine was flawed in that it did not correctly
map user priorities onto its intermittent matrix. The intermittent
matrix is used to group similarly traffic types for collapsing onto
available traffic classes.

Additionally with this fixed the 3 traffic class requirement can
be removed. Now we only require as many traffic classes as there
are traffic types. For example in the FCoE case with a best
effort and pfc traffic type the required minimum is 2.

Finally some additional comments were added and an example to
hopefully illustrate the solution.

Also I do realize that this can be done without using an
intermittent table and for that matter probably without even
the last table. And done in one shot. I believe this
implementation is clearer and easier to read. Although I would
welcome being corrected in the form of a patch submission.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>