lldptool: fix operation of -c option
If '-c' is entered with no argument list, then
lldptool should return all of the configurable elements
for the given TLV.  Current behavior requires that a
one or more configuration items be listed with '-c'.

The routines that get/set TLV settings in
lldp_mand_cmds.c loop through all the TLV
modules and invoke the get/set routines for
each TLV.  These routines are expected to
return 'cmd_not_applicable' if the cmd
is for a TLV not supported by the routine.
The get_arg_role() routine in lldp_vdp_cmds.c
was returning 'cmd_invalid' in such a way that
this scan was terminated prematurely.  The fix
is to move the code returning the error into
a part of the routine that is executed only after
the relevant VDP tlvid has been identified.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>