lldpad: fix segfault that occurs when config goes away
When the config directory is moved or deleted after
lldpad is running, a new interface coming up can result
in lldpad segfaulting.  This scenario can occur when
lldpad is run in initrd and interfaces come up after
chroot has occurred to the runtime system, but lldpad
has not yet been restarted for runtime.

A couple invocations of config_lookup() were not
checking for a NULL return value.

Signed-off-by: Eric Multanen <eric.w.multanen@intel.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>