lldpad: agent lookup must advance port
Setting the LLDP agent adminStatus does a port lookup by using the
interface name and agent type. When the interface name matches but
the agent type is not found the port _must_ be advanced.

We can not just break out of the loop to handle cases where 'all'
is set.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>