lldpad: mand: Use most correct chassisid/portid when possible
ChassisID and PortID TLVs are built once at init time. This is
intended to support sections 8.5.2 and 8.5.3 IEEE 802.AB 2009 spec.

"Chassis ID TLV and the chassis ID field value shall remain constant
 for all LLDPDUs while the MAC status parameter for the Port remains
 operational"

But because we read these TLVs out of the config file and shared
memory the user can never change these values for the life of the
config file. Also the user can not configure the subtype.

With this patch the user can now configure the subtype to be one
of the following lldpad supported types for chassid,

	CHASSIS_ID_MAC_ADDRESS
	CHASSIS_ID_NETWORK_ADDRESS
	CHASSIS_ID_INTERFACE_NAME

And one of the following portid types,

	PORT_ID_MAC_ADDRESS
	PORT_ID_NETWORK_ADDRESS
	PORT_ID_INTERFACE_NAME

The chassisID and portID will also now be rebuilt if either

	(a) link event occurs (MAC status goes non-operationalr[1])
	(b) SHM is removed and lldpad is restarted
	(c) user changes chassid or portid subtype

The shared memory (SHM) stores the MSAP (chassID + portID) and
is used by lldpad at init to allow the MSAP to be consistent
when transitioning from boot to runtime. To remove the shared
memory and effectively reset the MSAP at runtime the user can
run

	#lldpad -s
	#lldpad -d

or by toggling the link with administrator tools for example,

	#ip link set dev ethx down
	#ip link set dev ethx up

To set the subtype use the normal lldptool commmands structure,

	#lldptool -T -i ethx -V chassisid subtype=CHASSIS_ID_NETWORK_ADDRESS

[1] Defined in IEEE 802.1Q 6.6.2 Status parameters

The MAC_Operational parameter is TRUE if the entity providing the
service is capable of transmitting and receiving frames and its use
is permitted by management, i.e., MAC_Enabled is also TRUE. Its value
is otherwise FALSE. The value of this parameter is determined by the
specific MAC procedures, as specified in 6.7.

Reported-by: Shan Wei <shanwei88@gmail.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>
Tested-by: Shan Wei <shanwei88@gmail.com>
Tested-by: Ross Brattain <ross.b.brattain@intel.com>