lldpad: scan port can reference null agent head
Under stress conditions when lldpad misses a netlink messages and
the interface is being unregisteried it is possible for scan_port()
to try to iterate a NULL agent_head list. This occurs because
if_nameindex() returned interface names that lldpad has not yet
received a register netlink message for.

This bug was introduced with the agent handling code.

To resolve this the scan_port code is made more robust and will now
add any valid ports to the list of managed ports if detected through
the scan_port() path.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>