open-lldp: Add BASH completion files for lldptool and lldpad
These completion scripts will do general completion of
lldpad and lldptool options. The completion is fairly stupid
and doesn't differentiate between lldptool's commands and
options.

These scripts may or may not be added to the bash-completion
project and added to the distro's bash-completion packages.
'make install' will copy these scripts to the
/etc/bash_completion.d/ directory. However, they are not
usable unless the bash-completion package is installed.

Signed-off-by: Robert Love <robert.w.love@intel.com>
Tested-by: Ross Brattain <ross.b.brattain@intel.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>