lldpad: DCBX: DCB enabled by default in CEE mode on some kernels
When IEEE mode fails and falls back to CEE mode DCBX attributes
are advertised and DCB enabled by calling dont_advertise_dcbx_all().
As part of the ifup routine. This is setting the dcb_enable
field to true in the configuration file eventually resulting
in DCB hardware being enabled.

Normally this doesn't occur unless either a CEE DCBX TLV is
received or the user explicitly enables this mode. However,
on kernels < 2.6.38 IEEE is not supported and in this case
IEEE falls back to CEE mode immediately. This results in DCB
being enabled even though a peer tlv has not yet been
received and the user did not enable it.

Fix this logic by _only_ enabling, setting the dcb_enable
flag to "1", when DCB is explicitely enabled. This adds a
second option "2" which indicates DCB has not been manually
enabled and should only be enabled upon receiving a peer
TLV.

With this fix lldpad DCBX works as expected. IEEE mode is
the default and CEE mode is used if and only if a peer
CEE DCBX TLV is received or the user enables it through
the command line.

v2: Eric Multanen suggested this nicer version of the patch
    via private email. I did some cleanup and a few minor
    fixes to get to this version. This addresses the core
    issue that the 'dcb_enabled' field gets set rather than
    just moving it around and hiding it like my previous
    patch.

Suggested-by: Eric Multanen <eric.w.multanen@intel.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>