lldpad: Add peer pid to the VDP assoication
The pid of a VDP/VSI Association is now stored in the VSI profile with the
MAC/VLAN pair. The pid is changed when a different process sends the same
VSI profile information.
Add the Qos field to the settings.

Fixes a OOM problem in the function vdp_take_over_macvlans.

Signed-off-by: Thomas Richter <tmricht@linux.vnet.ibm.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>