lldpad: Avoid peer notification of switch dis-associate response.
When peer sends a SETLINK message to lldpad to start a VSI association
the peer queries lldpad for the status of the transaction. While this
is ongoing, do not send an unsolicited dis-associate originated by the switch.
The switch response is returned by the GETLINK netlink response.
Avoid entering VSI exit state too early otherwise the profile is removed
when the GETLINK response is constructed.

Signed-off-by: Thomas Richter <tmricht@linux.vnet.ibm.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>