lldpad: dcbtool app: enable and willing bits are global values
The Application data in the CEE DCBX TLV allows multiple APP
entries but uses a single set of enabled and willing bits. We
had been allowing users to configured these bits independently
per subtype and ended up reporting incorrect values through
dcbtool. And also had to pick one of the subtypes values to
send.

With this patch we allow users to set the enable and willing
bits but propagate the values to all APP entries.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>