lldpad: dcbx: turn down logging on dcb_fixup_pg
dcb_fixup_pg is logging messages as errors that occur when DCB is
disabled. Remove num_tc warning its not an error at all and is
indicating DCB is not enabled on the device yet. Then set the
warning returned from dcb_fixup_pg to DBG because this can be
helpful while debugging but normal users should not see this.

Suggested-by: Ross Brattain <ross.b.brattain@intel.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>