lldpad: regression: DCBX may fail to program hardware
I introduced a regression here,

commit 5e5af73d82598a35e3e3ab549de51ebd494ce945
Author: John Fastabend <john.r.fastabend@intel.com>
Date:   Tue Jan 10 08:42:10 2012 -0800

    lldpad: DCBX: resolve race with dropped IF_DORMANT msg

the patch added the operup field which is set to 'true' after
the APP and PFC tlvs become operational or the dormantTimer
expires. This is used to work around hardware that resets the
device while making DCB changes. For example changing the APP
TLV causes a link reset on Intel 82599 devices.

The patch also set the operup field to 'false' on link up
events. When transitioning from DOWN -> DORMANT the dormantTImer
starts and sets the operup field to 'false' to keep hardware
sets from occuring while the protocol negotiates and reaches
a steady state.

A regresssion occurs when the operup field is set to 'false'
on the transition from DORMANT -> UP. In this case the port
is already enabled and the dormantTimer has not been started.
The dormantTimer is not started and so the operup field gets
stuck 'false' and sets are not occuring in the normal path:

        /* apply all feature setting to the driver: linux only */
        tlvs = dcbx_data(device_name);
        if (tlvs && tlvs->operup) {
                LLDPAD_DBG("%s: %s: Managed DCB device coming online, program HW\n",
                            __func__, device_name);
                set_hw_all(device_name);
        }

This patch makes a change to only set the operup field to 'false'
when the dormantTimer is set and the port has not been enabled.
This should cover both transitions DOWN -> DORMANT -> UP which
occurs in most cases and the transition DOWN -> UP where the
dormant event was lost or never occured.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>