lldpad: Run VDP station state machine immediately
When the switch responds vdp_indicate() is called to reflect the
switch changes in the profile. It takes up to the next ack timeout
before the changes are processed.

This patch introduces a new function vdp_advance_sm() which is called
after the successful reception of a switch ECP packages. This function
advances the state machine immediately instead of waiting for the
next ack time out.

Signed-off-by: Thomas Richter <tmricht@linux.vnet.ibm.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>