lldpad: dcbx: app_cfg cfg file check expects two octet string
The get_persistent() routine used to read the dcbx section of the
config file expect the app_cfg field to be a two octet string. If
this is not true a sanity check fails and the routine uses default
DCBX values.

We could either relax the checks in get_persistent or fix the
cases where an int is written into the config file instead of a
two octet string.

This patch ensures the config file always uses a two byte octet.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>