lldpad: dcbx: dcb_user_priority overloads pgid field
The pgid field is used for bwgid in CIN mode and pgid in the
CEE case. This can cause unnecessary confusion especially when
the priority to priority group mapping is not 1:1. This may
result in a bug in drivers that implement the bandwidth
groups arbiters in CEE mode. Currently lldpad is passing
configurations at the driver which suggest a bwgid and uppct
that is inconsistent with the negotiated TLVs pgid:pgpct.

For example the configuration below might suggest the driver
can divide the 35% bandwidth on pgid 0 amongst all TCs except
TC 3. However because the up2tc mapping which is programmed
into the skb->priority to TC mapping is setup to map only
to TCs 0 and TCs 1 any credits on other TCs go unused.

# dcbtool gc eth3 pg
Command:        Get Config
Feature:        Priority Groups
Port:           eth3
Status:         Successful
Enable:         true
Advertise:      false
Willing:        false
up2tc:          0       0       0       1       0       0       0       0
pgpct:          35%     65%     0%      0%      0%      0%      0%      0%
pgid:           0       0       0       1       0       0       0       0
uppct:          15%     15%     14%     100%    14%     14%     14%     14%
pg strict:      0       0       0       0       0       0       0       0
num TC's:       8

Per DCB spec then unused credits are distributed amongst
the remaining TCs and the result may be TC0 and TC1 consuming
the extra credits in a device specific mechanism. This is
unlikely to match the expected 35%/65% bandwidth split.

This patch does some renaming to remove the overloading case.
By removing the tcmap and adding an explicit bwgid we can resolve
the above ambiguity.

The tcmap field is safe to remove because it was always being set
to the pgid values anyways (a follow up patch should restore the
tcmap functionality correctly if it is needed). By adding a bwgid
field the CIN implementation and CEE implementations can be setup
explicitly removing any confusion. Now CEE modes will map bandwidth
groups 1:1 with PG values. For now this seems to be the easiest
fix to the described problem and I do not anticipate this should
cause any problems with existing driver implementations.

Also note that the dcbtool uppct command becomes mostly unneeded
in CEE mode now and should be set to '100%' always.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>