lldpad: dcbx: map link strict to correct traffic class
When priority to traffic class map is not 1:1 link strict is not
being set correctly. This moves link strict attribute settings
so that the up2tc mapping is considered.

Note the link strict priority group will use the first free user
priority available. This means if the user puts a gap in the
priority groups for example using 0 and 2 but not 1. The link
strict priority will use traffic class 1.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>