lldpad: Add vdptest program
Add the vdptest program to test lldpad VSI profile handling.
This test program reads a configuration file or command line parameters
and sends the appropriate data via netlink interface to lldpad.
Results from lldpad are compared to expected values.
Abort on first mismatch.
The configuration file containing the VSI data is not part of this setup.
Its vendor and switch specific.

An man page is also added for this program.

Specify --enable-debug in the configure step to build it.

Signed-off-by: Thomas Richter <tmricht@linux.vnet.ibm.com>