lldpad: generate shared library liblldp_clif.so
We want to add dcb support to NetworkManager,
the function of this feature is similar to dcbtool,
so we need lldpad to export an clif shared library,
then we can use this library to send requests to and query results from
lldpad.

And also we make dcbtool/lldptool link against liblldp_clif.so.

changelog
V2: rename it with liblldp_clif.so
V3: set version of liblldp_clif.so

Signed-off-by: Weiping Pan <wpan@redhat.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>