lldp: dormantTimer via scan_port on netlink socket error
When a socket error is received possibly indicating a dropped
link event message. The event handler enters the scan port
path to verify we didn't miss adding/removing/updating port
status.

If we drop a dormant event it is possible that we may never
reintialize and start the dormantTimer. This can cause
problems with modules that depend on the dormant timer. For
example dcbx uses this timer to set the interfaces operstate.
If we transition into a dormant state and never configure
the dormant timer the module may never push the operstate into
IF_OPER_UP.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>