lldp_evb_h convert some members from pointers to structures
Convert evb_data member access from pointer to structure references.
The structure evb_data contains pointers to 3 EVB tlv's structures.
The pointers to structures are replaced by structures.
This avoids the allocation and de-allocation of the TLVs all the time and
the necessary error recovery in case the memory allocation fails.

Also get rid of the unpacked EVB TLV. This is only needed to convert an
internal EVB TLV representation to a packed EVB TLV. Stack memory will
be used instead.

Signed-off-by: Thomas Richter <tmricht@linux.vnet.ibm.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>